// app/src/app/core/api/api-client.ts
export class SaveRsvpRequest implements ISaveRsvpRequest {
  firstName?: string;
  lastName?: string;
  email?: string;

  constructor(data?: ISaveRsvpRequest) {
    if (data) {
      for (var property in data) {
        if (data.hasOwnProperty(property))
          (<any>this)[property] = (<any>data)[property];
      }
    }
  }

  init(_data?: any) {
    if (_data) {
      this.firstName = _data['firstName'];
      this.lastName = _data['lastName'];
      this.email = _data['email'];
    }
  }

  static fromJS(data: any): SaveRsvpRequest {
    data = typeof data === 'object' ? data : {};
    let result = new SaveRsvpRequest();
    result.init(data);
    return result;
  }

  toJSON(data?: any) {
    data = typeof data === 'object' ? data : {};
    data['firstName'] = this.firstName;
    data['lastName'] = this.lastName;
    data['email'] = this.email;
    return data;
  }
}

export interface ISaveRsvpRequest {
  firstName?: string;
  lastName?: string;
  email?: string;
}